FIXME: need to update NameConverterBase.getValidName() at some point
This commit moves the core NameConverter logic to a class in core that can be extended by NameConverter in ui.common. This removes dependencies on CorePlugin and eclipse for jmc-core and still migrates the functionality. The rest of ui.common.labelingrules has also been migrated to core in this commit.