wip: cjfr-core is now happy; rjmx and rjmx.services.jfr are mirgrate to core
TODO: migrate the tests, manage the permissions and api, and fix /application

For the classes that required Eclipse imports and/or dealt with persistence, there are now two copies of that class that extend a base abstract class. For example, RJMXSingletonBase is extended by an RJMXSingleton class in both core and org.openjdk.jmc.rjmx.ui, and this allows a core implementation that doesn't rely on Eclipse and an implementation for the ui that can handle RJMX Preferences.

Additionally, the RJMXPlugin code that required Eclipse imports have been migrated to the RJMXUIPlugin. Classes in rjmx.ui that were brought in and reference RJMXPlugin will need to be updated to use RJMXUIPlugin.