poc: remove eclipse-plugin from rjmx
Opening the export restrictions from rjmx and adding a require-bundle to
many of the application bundles allowed rjmx to drop the eclipse-plugin
packaging type.

uitests run as expected, and the rjmx test suite runs okay as long as
the tycho-surefire plugin is used - the maven-surefire is not an okay
substitute.